IKASAN-1283 - refactoring sftp sample to showcase sftp consumer builder together with @Value properties
