+++ ikasaneip/sample/spring-boot/sftp/src/main/resources/application.properties
@@ -35,6 +35,7 @@ sftp.consumer.remoteHost=localhost
 sftp.consumer.remotePort=22
 sftp.consumer.sourceDirectory=.
 sftp.consumer.filenamePattern=.*txt
+sftp.consumer.knownHosts=
 
 fileGenerator.consumer.cronExpression=0 0/1 * * * ?
