+++ ikasaneip/sample/spring-boot/sftp/src/test/resources/application.properties
@@ -40,6 +40,7 @@ sftp.consumer.remoteHost=localhost
 sftp.consumer.remotePort=22999
 sftp.consumer.sourceDirectory=.
 sftp.consumer.filenamePattern=.*txt
+sftp.consumer.knownHosts=
 
 fileGenerator.consumer.cronExpression=0/10 * * * * ?
